% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aud_cpi}
\alias{aud_cpi}
\title{Australian CPI Data (1948-2024)}
\format{
A data frame with 77 rows and 2 columns:
\describe{
\item{year}{Calendar year (integer)}
\item{index}{CPI index value (numeric, base 2020 = 100)}
}
}
\usage{
aud_cpi
}
\description{
Annual average Consumer Price Index for Australia.
Index base: 2020 = 100. Source: Australian Bureau of Statistics (ABS).
}
\keyword{datasets}
