% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eur_cpi}
\alias{eur_cpi}
\title{Euro Area CPI Data (1960-2024)}
\format{
A data frame with 65 rows and 2 columns:
\describe{
\item{year}{Calendar year (integer)}
\item{index}{CPI index value (numeric, base 2020 = 100)}
}
}
\usage{
eur_cpi
}
\description{
Annual average Harmonised Index of Consumer Prices (HICP) for the Euro area.
Pre-1999 values are a GDP-weighted composite proxy of major Eurozone economies.
Index base: 2020 = 100. Source: Eurostat.
}
\keyword{datasets}
