% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflateR-package.R
\docType{package}
\name{inflateR-package}
\alias{inflateR}
\alias{inflateR-package}
\title{inflateR: Inflation Adjustment for Historical Currency Values}
\description{
Convert historical monetary values into their present-day equivalents using bundled CPI (Consumer Price Index) and GDP deflator data sourced from the World Bank Development Indicators. Supports British pounds (GBP), Australian dollars (AUD), US dollars (USD), Euro (EUR), Canadian dollars (CAD), Japanese yen (JPY), Chinese yuan (CNY), Swiss francs (CHF), New Zealand dollars (NZD), Indian rupees (INR), South Korean won (KRW), Brazilian reais (BRL), and Norwegian krone (NOK). Currency codes and country names are both accepted as input.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://charlescoverdale.github.io/inflateR/}
  \item \url{https://github.com/charlescoverdale/inflateR}
  \item Report bugs at \url{https://github.com/charlescoverdale/inflateR/issues}
}

}
\author{
\strong{Maintainer}: Charles Coverdale \email{charlesfcoverdale@gmail.com}

}
\concept{CPI}
\concept{GDP deflator}
\concept{cost of living}
\concept{currency adjustment}
\concept{inflation}
\concept{price index}
\concept{purchasing power}
\concept{real terms}
\keyword{internal}
